Add Hotline Webring links
Hotline Webring is the premier webring of our time. It offers a place
for like-minded sites to join together, holding hands through `next` and
`previous` links for all visitors to see. Becoming a member is truly an
honor.

I took it upon myself to bestow this honor in the following way:

When one hovers over the audio at the bottom of the page, the audio
disappears and the HLWR links appear. A side effect is that instead of
the hover trigger being on the audio element, it's now on its parent,
the .audio-container element. This is to make sure that the hover event
is respected, even when the hovering over the hotline webring links.
Otherwise they try and flip back and forth all crazy-like.